requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-extensions==3.0.9
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tabulate==0.8.7
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
